app/sentiment-analysis/requirements.txt
nltk==3.7
pandas==1.4.1
tensorflow==2.8.0
tensorflow-datasets==4.5.2
matplotlib==3.5.1
tweepy==4.8.0
requests==2.27.1
pymongo==4.0.1
python-dotenv==0.19.2
wordcloud==1.8.1